Move Maya payment integration from microservice to main API
