+++ api/.env.example
@@ -91,10 +91,9 @@ ANTHROPIC_API_KEY=
 ANTHROPIC_MODEL=claude-3-5-sonnet-latest
 ANTHROPIC_BASE_URL=https://api.anthropic.com
 
-# Payments microservice integration
-PAYMENTS_SERVICE_BASE_URL=http://localhost:8001/api
-PAYMENTS_SERVICE_INTERNAL_TOKEN=
-PAYMENTS_SERVICE_TIMEOUT=15
-
-# Internal callback auth token used by payments microservice -> main API
-PAYMENTS_CALLBACK_TOKEN=
+# Maya Checkout configuration (handled directly by main API)
+MAYA_BASE_URL=https://pg-sandbox.paymaya.com
+MAYA_PUBLIC_KEY=
+MAYA_SECRET_KEY=
+MAYA_WEBHOOK_SIGNATURE_KEY=
+MAYA_TIMEOUT=20
