+++ api/config/payments.php
@@ -1,12 +1,11 @@
 <?php
 
 return [
-    'service' => [
-        'base_url' => env('PAYMENTS_SERVICE_BASE_URL', 'http://localhost:8001/api'),
-        'internal_token' => env('PAYMENTS_SERVICE_INTERNAL_TOKEN'),
-        'timeout' => (int) env('PAYMENTS_SERVICE_TIMEOUT', 15),
-    ],
-    'callback' => [
-        'token' => env('PAYMENTS_CALLBACK_TOKEN'),
+    'maya' => [
+        'base_url' => env('MAYA_BASE_URL', 'https://pg-sandbox.paymaya.com'),
+        'public_key' => env('MAYA_PUBLIC_KEY'),
+        'secret_key' => env('MAYA_SECRET_KEY'),
+        'webhook_signature_key' => env('MAYA_WEBHOOK_SIGNATURE_KEY'),
+        'timeout' => (int) env('MAYA_TIMEOUT', 20),
     ],
 ];
