+++ microservices/payments/.env.example
@@ -18,17 +18,7 @@ DB_DATABASE=payments
 DB_USERNAME=payments_user
 DB_PASSWORD=payments_password
 
-# Internal token required for main API -> payments service calls
-PAYMENTS_INTERNAL_TOKEN=
-
-# Main API callback (payments service -> main API)
-PAYMENTS_CALLBACK_URL=
-PAYMENTS_CALLBACK_TOKEN=
-PAYMENTS_CALLBACK_TIMEOUT=10
-
-# Maya Checkout configuration
-MAYA_BASE_URL=https://pg-sandbox.paymaya.com
-MAYA_PUBLIC_KEY=
-MAYA_SECRET_KEY=
-MAYA_WEBHOOK_SIGNATURE_KEY=
-MAYA_TIMEOUT=20
+# Payment provider (e.g. Stripe) – add when integrating
+# STRIPE_KEY=
+# STRIPE_SECRET=
+# STRIPE_WEBHOOK_SECRET=
