--- microservices/payments/config/payments.php
@@ -1,19 +1,0 @@
-<?php
-
-return [
-    'internal' => [
-        'token' => env('PAYMENTS_INTERNAL_TOKEN'),
-    ],
-    'callback' => [
-        'url' => env('PAYMENTS_CALLBACK_URL'),
-        'token' => env('PAYMENTS_CALLBACK_TOKEN'),
-        'timeout' => (int) env('PAYMENTS_CALLBACK_TIMEOUT', 10),
-    ],
-    'maya' => [
-        'base_url' => env('MAYA_BASE_URL', 'https://pg-sandbox.paymaya.com'),
-        'public_key' => env('MAYA_PUBLIC_KEY'),
-        'secret_key' => env('MAYA_SECRET_KEY'),
-        'webhook_signature_key' => env('MAYA_WEBHOOK_SIGNATURE_KEY'),
-        'timeout' => (int) env('MAYA_TIMEOUT', 20),
-    ],
-];
