+++ microservices/payments/routes/api.php
@@ -9,13 +9,7 @@ Route::get('/', function () {
 
 // Payment service routes (main API will call these internally)
 Route::prefix('v1')->group(function () {
-    Route::middleware('internal.token')->group(function () {
-        Route::post('/charges', [PaymentController::class, 'createCharge']);
-        Route::get('/charges/{id}', [PaymentController::class, 'showCharge']);
-    });
-
-    // Public webhook endpoint called by Maya.
-    Route::post('/webhooks/maya', [PaymentController::class, 'mayaWebhook']);
-    // Backward compatible alias.
+    Route::post('/charges', [PaymentController::class, 'createCharge']);
+    Route::get('/charges/{id}', [PaymentController::class, 'showCharge']);
     Route::post('/webhooks/stripe', [PaymentController::class, 'stripeWebhook']);
 });
